Part of #416 
Optimizes the proof of work:

## Single hash benchmark:
```
BenchmarkPowHash-20    	 2889742	       394.6 ns/op	     152 B/op	       3 allocs/op
↓
BenchmarkPowHash-20    	 3915529	       302.4 ns/op	       0 B/op	       0 allocs/op
```

## Finding PoW benchmark:
```
BenchmarkFindSubmitPowNonce/difficulty=20-20                   2         625033028 ns/op        243352568 B/op   4802780 allocs/op
↓
BenchmarkFindSubmitPowNonce/difficulty=20-20                   3         348184603 ns/op              416 B/op         6 allocs/op
```